Refactor: move JSON page rendering to a separate project
The fitlayout-render-json project serves as a base for the
fitlayout-render-puppeteer and fitlayout-render-playwright modules.
