+++ bom/pom.xml
@@ -35,7 +35,17 @@
 			</dependency>
 			<dependency>
 				<groupId>cz.vutbr.fit.layout</groupId>
+				<artifactId>fitlayout-render-json</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>cz.vutbr.fit.layout</groupId>
 				<artifactId>fitlayout-render-puppeteer</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>cz.vutbr.fit.layout</groupId>
+				<artifactId>fitlayout-render-playwright</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
