+++ fitlayout-render-json/pom.xml
@@ -8,11 +8,11 @@
 		<version>2.0.4-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>fitlayout-render-puppeteer</artifactId>
+	<artifactId>fitlayout-render-json</artifactId>
 
-	<name>FitLayout Puppeteer Client</name>
+	<name>FitLayout JSON page renderer</name>
 	<description>
-		FitLayout - Puppeteer/Chromium renderer client
+		FitLayout - Generic JSON-based page description renderer
 	</description>
 
 	<build>
@@ -25,12 +25,17 @@
 			<artifactId>fitlayout-core</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>cz.vutbr.fit.layout</groupId>
-			<artifactId>fitlayout-render-json</artifactId>
-		</dependency>
-		<dependency>
 			<groupId>org.slf4j</groupId>
 			<artifactId>slf4j-api</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>com.google.code.gson</groupId>
+			<artifactId>gson</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>net.sf.cssbox</groupId>
+			<artifactId>jstyleparser</artifactId>
+			<version>4.0.0</version>
 		</dependency>
 	</dependencies>
