+++ fitlayout-render-puppeteer/src/main/java/cz/vutbr/fit/layout/puppeteer/impl/BoxTreeBuilder.java
@@ -7,14 +7,11 @@ package cz.vutbr.fit.layout.puppeteer.impl;
 
 import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
-import java.util.Base64;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
@@ -24,20 +21,16 @@ import org.slf4j.LoggerFactory;
 
 import com.google.gson.Gson;
 
-import cz.vutbr.fit.layout.impl.BaseBoxTreeBuilder;
-import cz.vutbr.fit.layout.model.Box;
-import cz.vutbr.fit.layout.model.Color;
-import cz.vutbr.fit.layout.model.Metadata;
-import cz.vutbr.fit.layout.model.Page;
-import cz.vutbr.fit.layout.puppeteer.parser.InputFile;
-import cz.vutbr.fit.layout.puppeteer.parser.MetadataDef;
-import cz.vutbr.fit.layout.puppeteer.parser.PageInfo;
+import cz.vutbr.fit.layout.json.impl.JSONBoxTreeBuilder;
+import cz.vutbr.fit.layout.json.impl.StreamConsumer;
+import cz.vutbr.fit.layout.json.parser.InputFile;
 
 /**
- * 
+ * A JSON-based builder that uses the fitlayout-puppeteer backend for obtaining the source JSON page desription.
+ *  
  * @author burgetr
  */
-public class BoxTreeBuilder extends BaseBoxTreeBuilder
+public class BoxTreeBuilder extends JSONBoxTreeBuilder
 {
     private static final String PROP_BACKEND = "fitlayout.puppeteer.backend";
     private static final String PROP_WORKDIR = "fitlayout.puppeteer.workdir";
@@ -47,12 +40,6 @@ public class BoxTreeBuilder extends BaseBoxTreeBuilder
     
     public static final String DEFAULT_FONT_FAMILY = "sans-serif";
     public static final float DEFAULT_FONT_SIZE = 12;
-    
-    /** Input JSON representation */
-    private InputFile inputFile;
-    
-    /** The resulting page */
-    private PageImpl page;
     
     /** Browser window width for rendering */
     private int width;
@@ -90,96 +77,6 @@ public class BoxTreeBuilder extends BaseBoxTreeBuilder
     {
         this.includeScreenshot = includeScreenshot;
     }
-
-    public void parse(String urlstring) throws MalformedURLException, IOException, InterruptedException
-    {
-        urlstring = urlstring.trim();
-        if (urlstring.startsWith("http:") ||
-            urlstring.startsWith("https:") ||
-            urlstring.startsWith("ftp:") ||
-            urlstring.startsWith("file:"))
-        {
-            parse(new URL(urlstring));
-        }
-        else if (urlstring.startsWith("json:"))
-        {
-            parseJSON(urlstring.substring(5));
-        }
-        else
-            throw new MalformedURLException("Unsupported protocol in " + urlstring);
-    }
-    
-    public void parse(URL url) throws IOException, InterruptedException
-    {
-        inputFile = invokeRenderer(url);
-        if (inputFile != null && inputFile.getError() == null)
-        {
-            if (inputFile.getStatus() >= 200 && inputFile.getStatus() < 300)
-                parseInputFile(inputFile, url);
-            else
-                throw new IOException("HTTP status: " + inputFile.getStatus() + " " + inputFile.getStatusText());
-        }
-        else if (inputFile.getError() != null)
-            throw new IOException(inputFile.getError());
-        else
-            throw new IOException("Backend execution failed");
-    }
-    
-    public void parseJSON(String path) throws IOException
-    {
-        inputFile = loadJSON(path);
-        URL url;
-        try {
-            if (inputFile.page != null && inputFile.page.url != null)
-                url = new URL(inputFile.page.url);
-            else
-                url = new URL("http://url.not.available");
-        } catch (MalformedURLException e) {
-            url = new URL("http://url.not.available");
-        }
-        parseInputFile(inputFile, url);
-    }
-    
-    protected void parseInputFile(InputFile input, URL url) throws IOException
-    {
-        inputFile = input;
-
-        //create the page
-        PageInfo pInfo = inputFile.getPage();
-        page = new PageImpl(url);
-        page.setTitle(pInfo.getTitle());
-        page.setWidth(Math.round(pInfo.getWidth()));
-        page.setHeight(Math.round(pInfo.getHeight()));
-        if (input.getScreenshot() != null)
-        {
-            try {
-                byte[] image = Base64.getDecoder().decode(input.getScreenshot());
-                page.setPngImage(image);
-            } catch (IllegalArgumentException e) {
-                log.error("Couldn't decode a base64 screenshot: {}", e.getMessage());
-            }
-        }
-        
-        //create the box tree
-        BoxList boxlist = new BoxList(inputFile);
-        Box root = buildTree(boxlist.getVisibleBoxes(), Color.WHITE);
-        page.setRoot(root);
-        
-        //copy the metadata (if provided)
-        if (input.getMetadata() != null)
-        {
-            List<Metadata> data = new ArrayList<>();
-            for (MetadataDef item : input.getMetadata())
-                data.add(item);
-            page.setMetadata(data);
-        }
-    }
-    
-    @Override
-    public Page getPage()
-    {
-        return page;
-    }
     
     //==================================================================================
     
@@ -190,7 +87,7 @@ public class BoxTreeBuilder extends BaseBoxTreeBuilder
      * @throws IOException
      * @throws InterruptedException
      */
-    private InputFile invokeRenderer(URL url) throws IOException, InterruptedException
+    protected InputFile invokeRenderer(URL url) throws IOException, InterruptedException
     {
         String rendererPath = System.getProperty(PROP_BACKEND); //puppeteer backend project path
         if (rendererPath == null)
@@ -280,21 +177,6 @@ public class BoxTreeBuilder extends BaseBoxTreeBuilder
             throw new IOException(msg);
         }
         InputFile file = (InputFile) outConsumer.getResult();
-        return file;
-    }
-    
-    /**
-     * Parses a local file produced by the backend.
-     * @param path the path to the file to parse
-     * @return the parsed file
-     * @throws IOException
-     */
-    private InputFile loadJSON(String path) throws IOException
-    {
-        FileReader fin = new FileReader(path);
-        Gson gson = new Gson();
-        InputFile file = gson.fromJson(fin, InputFile.class);
-        fin.close();
         return file;
     }
